swap whitelist
