+++ x/ammswap/client/rest/query.go
@@ -15,12 +15,9 @@ import (
 
 func registerQueryRoutes(cliCtx context.CLIContext, r *mux.Router) {
 	r = r.PathPrefix("/swap").Subrouter()
-	r.HandleFunc("/tokens", swapTokensHandler(cliCtx)).Methods("GET")
 	r.HandleFunc("/token_pair/{name}", querySwapTokenPairHandler(cliCtx)).Methods("GET")
-	r.HandleFunc("/token_pairs", querySwapTokenPairsHandler(cliCtx)).Methods("GET")
 	r.HandleFunc("/params", queryParamsHandler(cliCtx)).Methods("GET")
 	r.HandleFunc("/liquidity/add_quote/{token}", swapAddQuoteHandler(cliCtx)).Methods("GET")
-	r.HandleFunc("/liquidity/histories", swapLiquidityHistoriesHandler(cliCtx)).Methods("GET")
 	r.HandleFunc("/liquidity/remove_quote/{token_pair}", queryRedeemableAssetsHandler(cliCtx)).Methods("GET")
 	r.HandleFunc("/quote/{token}", swapQuoteHandler(cliCtx)).Methods("GET")
 }
@@ -36,21 +33,6 @@ func querySwapTokenPairHandler(cliContext context.CLIContext) func(http.Response
 			return
 		}
 		rest.PostProcessResponse(w, cliContext, res)
-	}
-
-}
-
-func querySwapTokenPairsHandler(cliContext context.CLIContext) func(http.ResponseWriter, *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-
-		res, _, err := cliContext.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QuerySwapTokenPairs), nil)
-		if err != nil {
-			sdkErr := common.ParseSDKError(err.Error())
-			common.HandleErrorMsg(w, cliContext, sdkErr.Code, sdkErr.Message)
-			return
-		}
-
-		formatAndReturnResult(w, cliContext, res)
 	}
 
 }
@@ -99,30 +81,6 @@ func formatAndReturnResult(w http.ResponseWriter, cliContext context.CLIContext,
 	rest.PostProcessResponse(w, cliContext, resultJson)
 }
 
-func swapTokensHandler(cliCtx context.CLIContext) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		businessType := r.URL.Query().Get("business_type")
-		address := r.URL.Query().Get("address")
-		baseTokenName := r.URL.Query().Get("base_token")
-
-		params := types.NewQuerySwapTokensParams(businessType, address, baseTokenName)
-		bz, err := cliCtx.Codec.MarshalJSON(params)
-		if err != nil {
-			common.HandleErrorMsg(w, cliCtx, common.CodeMarshalJSONFailed, err.Error())
-			return
-		}
-
-		res, _, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QuerySwapTokens), bz)
-		if err != nil {
-			sdkErr := common.ParseSDKError(err.Error())
-			common.HandleErrorMsg(w, cliCtx, sdkErr.Code, sdkErr.Message)
-			return
-		}
-
-		rest.PostProcessResponse(w, cliCtx, res)
-	}
-}
-
 func swapQuoteHandler(cliCtx context.CLIContext) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
@@ -137,29 +95,6 @@ func swapQuoteHandler(cliCtx context.CLIContext) http.HandlerFunc {
 		}
 
 		res, _, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QuerySwapQuoteInfo), bz)
-		if err != nil {
-			sdkErr := common.ParseSDKError(err.Error())
-			common.HandleErrorMsg(w, cliCtx, sdkErr.Code, sdkErr.Message)
-			return
-		}
-
-		rest.PostProcessResponse(w, cliCtx, res)
-	}
-}
-
-func swapLiquidityHistoriesHandler(cliCtx context.CLIContext) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		address := r.URL.Query().Get("address")
-		tokenPairName := r.URL.Query().Get("token_pair_name")
-
-		params := types.NewQuerySwapLiquidityInfoParams(address, tokenPairName)
-		bz, err := cliCtx.Codec.MarshalJSON(params)
-		if err != nil {
-			common.HandleErrorMsg(w, cliCtx, common.CodeMarshalJSONFailed, err.Error())
-			return
-		}
-
-		res, _, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", types.QuerierRoute, types.QuerySwapLiquidityHistories), bz)
 		if err != nil {
 			sdkErr := common.ParseSDKError(err.Error())
 			common.HandleErrorMsg(w, cliCtx, sdkErr.Code, sdkErr.Message)
