+++ x/backend/keeper/querier.go
@@ -74,6 +74,12 @@ func NewQuerier(keeper Keeper) sdk.Querier {
 
 		case types.QuerySwapWatchlist:
 			res, err = querySwapWatchlist(ctx, req, keeper)
+		case types.QuerySwapTokens:
+			res, err = querySwapTokens(ctx, req, keeper)
+		case types.QuerySwapTokenPairs:
+			res, err = querySwapTokenPairs(ctx, path[1:], req, keeper)
+		case types.QuerySwapLiquidityHistories:
+			res, err = querySwapLiquidityHistories(ctx, req, keeper)
 		case types.QueryTickerListV2:
 			if keeper.Config.EnableMktCompute {
 				res, err = queryTickerListV2(ctx, path[1:], req, keeper)
