+++ x/backend/orm/swap.go
@@ -36,3 +36,36 @@ func (orm *ORM) GetSwapInfo(startTime int64) []types.SwapInfo {
 	query.Order("timestamp asc").Find(&swapInfos)
 	return swapInfos
 }
+
+// AddSwapWhitelist insert swap whitelist to db
+func (orm *ORM) AddSwapWhitelist(swapWhitelists []*types.SwapWhitelist) (addedCnt int, err error) {
+	orm.singleEntryLock.Lock()
+	defer orm.singleEntryLock.Unlock()
+
+	tx := orm.db.Begin()
+	defer orm.deferRollbackTx(tx, err)
+	cnt := 0
+
+	for _, swapWhitelist := range swapWhitelists {
+		if swapWhitelist != nil {
+			ret := tx.Create(swapWhitelist)
+			if ret.Error != nil {
+				return cnt, ret.Error
+			} else {
+				cnt++
+			}
+		}
+	}
+
+	tx.Commit()
+	return cnt, nil
+}
+
+// nolint
+func (orm *ORM) GetSwapWhitelist() []types.SwapWhitelist {
+	var swapWhitelist []types.SwapWhitelist
+	query := orm.db.Model(types.SwapWhitelist{}).Where("deleted = false")
+
+	query.Order("timestamp asc").Find(&swapWhitelist)
+	return swapWhitelist
+}
